Add new event classes for gameplay mechanics including GoldFromHeaven, Block, MakeKey, Teleport, and StMichael
